+++ backend/src/common/result.entity.ts
